Consolidate delay UGens
