--- supriya/ugens/Delay1.py
@@ -1,25 +1,0 @@
-import collections
-
-from supriya import CalculationRate
-from supriya.synthdefs import PureUGen
-
-
-class Delay1(PureUGen):
-    """
-    A one-sample delay line unit generator.
-
-    ::
-
-        >>> source = supriya.ugens.In.ar(bus=0)
-        >>> supriya.ugens.Delay1.ar(source=source)
-        Delay1.ar()
-
-    """
-
-    ### CLASS VARIABLES ###
-
-    __documentation_section__ = "Delay UGens"
-
-    _ordered_input_names = collections.OrderedDict([("source", None)])
-
-    _valid_calculation_rates = (CalculationRate.AUDIO, CalculationRate.CONTROL)
